change type of 'id' from Long to int; replace class IdNotFoundException with ResourceNotFoundException and add ExceptionHandler; modify methods in service and web accordingly

+++ WWR-API/src/main/java/com/infoshare/alpha/wwr/dao/FacilityDao.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public interface FacilityDao {
 
-    Optional<Facility> findById(Long id);
+    Optional<Facility> findById(int id);
 
     List<Facility> findAll();
 }
